refactor(update): own notification eligibility
