+++ internal/update/notification_test.go
@@ -1,4 +1,4 @@
-package cmd
+package update
 
 import (
 	"os"
@@ -9,11 +9,9 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"github.com/shopware/shopware-cli/internal/update"
 )
 
-func TestShouldNotifySuppressesRecentHomebrewRelease(t *testing.T) {
+func TestShouldNotify(t *testing.T) {
 	if runtime.GOOS == "windows" {
 		t.Skip("fake brew shell script requires a unix shell")
 	}
@@ -23,11 +21,11 @@ func TestShouldNotifySuppressesRecentHomebrewRelease(t *testing.T) {
 	require.NoError(t, os.WriteFile(filepath.Join(binDir, "brew"), []byte("#!/bin/sh\necho "+prefix+"\n"), 0o755))
 	t.Setenv("PATH", binDir)
 
-	recent := &update.ReleaseInfo{Version: "99.0.0", PublishedAt: time.Now()}
-	old := &update.ReleaseInfo{Version: "99.0.0", PublishedAt: time.Now().Add(-48 * time.Hour)}
+	recent := &ReleaseInfo{Version: "99.0.0", PublishedAt: time.Now()}
+	old := &ReleaseInfo{Version: "99.0.0", PublishedAt: time.Now().Add(-48 * time.Hour)}
 	homebrewBinary := filepath.Join(prefix, "bin", "shopware-cli")
 
-	assert.False(t, shouldNotify(recent, homebrewBinary))
-	assert.True(t, shouldNotify(recent, "/usr/local/other/shopware-cli"))
-	assert.True(t, shouldNotify(old, homebrewBinary))
+	assert.False(t, ShouldNotify(recent, homebrewBinary))
+	assert.True(t, ShouldNotify(recent, "/usr/local/other/shopware-cli"))
+	assert.True(t, ShouldNotify(old, homebrewBinary))
 }
